remove PAC and shape and move SWM
