--- neurodsp/shape/__init__.py
@@ -1,4 +1,0 @@
-from .cyclefeatures import features_by_cycle, plot_burst_detect_params
-from .cyclepoints import find_extrema, find_zerox
-from .phase import extrema_interpolated_phase
-from .swm import sliding_window_matching
