+++ neurodsp/swm.py
